+++ src/main/java/org/mju_likelion/festival/auth/util/key/RsaKeyStrategy.java
@@ -1,4 +1,4 @@
-package org.mju_likelion.festival.auth.domain;
+package org.mju_likelion.festival.auth.util.key;
 
 import static org.mju_likelion.festival.common.exception.type.ErrorType.INVALID_RSA_KEY_STRATEGY_ERROR;
