+++ src/main/java/org/mju_likelion/festival/auth/util/key/manager/RedisRsaKeyManager.java
@@ -4,8 +4,8 @@ import static org.mju_likelion.festival.common.exception.type.ErrorType.CREDENTI
 
 import java.util.UUID;
 import lombok.AllArgsConstructor;
-import org.mju_likelion.festival.auth.domain.RsaKey;
-import org.mju_likelion.festival.auth.domain.RsaKeyStrategy;
+import org.mju_likelion.festival.auth.util.key.RsaKey;
+import org.mju_likelion.festival.auth.util.key.RsaKeyStrategy;
 import org.mju_likelion.festival.auth.util.key.RsaKeyUtil;
 import org.mju_likelion.festival.common.exception.NotFoundException;
 import org.mju_likelion.festival.common.util.redis.RedisUtil;
